Update DownFriday-DownMonday.R
Rendered the analysis as percentage of share price rather than dollars per share. This somewhat weakened the measured effect size, which is now less than 1% of share price in all 28 model specifications and slightly negative in 2/28 model specifications. 

Effect size remains smallest on the longest time frames. I thought that using percentages rather than dollars would change that, but it remains as true in percentage terms as it was in dollar terms. This raises the real possibility that the recent strong performance of DF+DM is a historical anomaly that disappears on larger time scales.